Add doc.go file
Also make the default values for the Client configuration as constants to make it more apparent what they are. 
Broke out the connect/disconnect logic from the client.go file to connect.go.
